+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 * Updates for upcoming Platform Networking feature, including new SyncSocketProvider class. ([PR #6096](https://github.com/realm/realm-core/pull/6096))
 * Updated namespaces for files moved to realm/sync/network ([PR #6109](https://github.com/realm/realm-core/pull/6109))
 * Replace util::network::Trigger with a Sync client custom trigger. ([PR #6121](https://github.com/realm/realm-core/pull/6121))
+* Create DefaultSyncSocket class ([PR #6116](https://github.com/realm/realm-core/pull/6116))
 
 ----------------------------------------------